moving Profile data fetch to BE
