+++ components/shared/Loading/Loading.tsx
@@ -1,0 +1,7 @@
+import { Center, Spinner } from "@chakra-ui/react";
+
+export const Loading = () => (
+  <Center h={"100px"}>
+    <Spinner color="white" />
+  </Center>
+);
