+++ components/shared/SignInGate/SignInGate.tsx
@@ -2,12 +2,13 @@ import { Box } from "@chakra-ui/react";
 import { useUserSession } from "components/context/UserSessionContext";
 import React from "react";
 import { EmailAuthModal } from "../EmailAuthModal";
+import { Loading } from "../Loading";
 
 export const SignInGate = ({ children }: { children: React.ReactElement }) => {
   const { isLoading, user } = useUserSession();
 
   if (isLoading) {
-    return <h1>Loading...</h1>;
+    return <Loading />;
   }
 
   return !!user ? (
